feat: Major update to data_loader
Instead of retunring the binary masks alone we also retunr a skeletonized version which describes the cracks with 9 parameters and a a vectore with keypoints from which we can reconstruct the original damage mask.
The idea is to reduce the dimensionality of the damage mask but still be able to reconstruct the original mask.

This version works partially. It produces descriptors which can be reconstructud with about 85% accuracy. This needs some more work but is a good baseline.