FEAT: This version implements the MMVAE form Unity by Diversity using only two modalities. The Spectograms computed from the Raw time sseries data as well as the descriptors computed from the masks.
These computations use a lot of RAM and thus I use Lazy data loading and batches. That causes this version to be very slow.
There are also quite a bit of warning massages (C++) which are partially surpressed and come from the Docker Image pulled to run TF with GPU lavrage.